--- a/src/components/DayPickerRangeController.jsx
+++ b/src/components/DayPickerRangeController.jsx
@@ -1006,17 +1006,16 @@ export default class DayPickerRangeController extends React.PureComponent {
         Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
       ));
 
-      updatedDaysAfterAddition = monthsToUpdate.reduce((days, monthIso) => {
+      updatedDaysAfterAddition = monthsToUpdate.reduce((acc, monthIso) => {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
         const modifiers = new Set(month[iso]);
         modifiers.add(modifier);
-        return {
-          ...days,
+        return Object.assign(acc, {
           [monthIso]: {
             ...month,
             [iso]: modifiers,
           },
-        };
+        });
       }, updatedDaysAfterAddition);
     } else {
       const monthIso = toISOMonthString(day);
@@ -1071,17 +1070,16 @@ export default class DayPickerRangeController extends React.PureComponent {
         Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
       ));
 
-      updatedDaysAfterDeletion = monthsToUpdate.reduce((days, monthIso) => {
+      updatedDaysAfterDeletion = monthsToUpdate.reduce((acc, monthIso) => {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
         const modifiers = new Set(month[iso]);
         modifiers.delete(modifier);
-        return {
-          ...days,
+        return Object.assign(acc, {
           [monthIso]: {
             ...month,
             [iso]: modifiers,
           },
-        };
+        });
       }, updatedDaysAfterDeletion);
     } else {
       const monthIso = toISOMonthString(day);