--- a/src/components/DayPickerRangeController.jsx
+++ b/src/components/DayPickerRangeController.jsx
@@ -1009,12 +1009,14 @@ export default class DayPickerRangeController extends React.PureComponent {
       updatedDaysAfterAddition = monthsToUpdate.reduce((acc, monthIso) => {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
 
-        const modifiers = new Set(month[iso]);
-        modifiers.add(modifier);
-        acc[monthIso] = {
-          ...month,
-          [iso]: modifiers,
-        };
+        if (!month[iso] || !month[iso].has(modifier)) {
+          const modifiers = new Set(month[iso]);
+          modifiers.add(modifier);
+          acc[monthIso] = {
+            ...month,
+            [iso]: modifiers,
+          };
+        }
 
         return acc;
       }, updatedDaysAfterAddition);
@@ -1022,12 +1024,14 @@ export default class DayPickerRangeController extends React.PureComponent {
       const monthIso = toISOMonthString(day);
       const month = updatedDays[monthIso] || visibleDays[monthIso];
 
-      const modifiers = new Set(month[iso]);
-      modifiers.add(modifier);
-      updatedDaysAfterAddition[monthIso] = {
-        ...month,
-        [iso]: modifiers,
-      };
+      if (!month[iso] || !month[iso].has(modifier)) {
+        const modifiers = new Set(month[iso]);
+        modifiers.add(modifier);
+        updatedDaysAfterAddition[monthIso] = {
+          ...month,
+          [iso]: modifiers,
+        };
+      }
     }
 
     return updatedDaysAfterAddition;
@@ -1070,24 +1074,30 @@ export default class DayPickerRangeController extends React.PureComponent {
 
       updatedDaysAfterDeletion = monthsToUpdate.reduce((acc, monthIso) => {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
-        const modifiers = new Set(month[iso]);
-        modifiers.delete(modifier);
-        acc[monthIso] = {
-          ...month,
-          [iso]: modifiers,
-        };
+
+        if (month[iso] && month[iso].has(modifier)) {
+          const modifiers = new Set(month[iso]);
+          modifiers.delete(modifier);
+          acc[monthIso] = {
+            ...month,
+            [iso]: modifiers,
+          };
+        }
+
         return acc;
       }, updatedDaysAfterDeletion);
     } else {
       const monthIso = toISOMonthString(day);
       const month = updatedDays[monthIso] || visibleDays[monthIso];
 
-      const modifiers = new Set(month[iso]);
-      modifiers.delete(modifier);
-      updatedDaysAfterDeletion[monthIso] = {
-        ...month,
-        [iso]: modifiers,
-      };
+      if (month[iso] && month[iso].has(modifier)) {
+        const modifiers = new Set(month[iso]);
+        modifiers.delete(modifier);
+        updatedDaysAfterDeletion[monthIso] = {
+          ...month,
+          [iso]: modifiers,
+        };
+      }
     }
 
     return updatedDaysAfterDeletion;