--- a/src/components/DayPickerSingleDateController.jsx
+++ b/src/components/DayPickerSingleDateController.jsx
@@ -535,14 +535,15 @@ export default class DayPickerSingleDateController extends React.PureComponent {
 
       updatedDaysAfterAddition = monthsToUpdate.reduce((acc, monthIso) => {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
+
         const modifiers = new Set(month[iso]);
         modifiers.add(modifier);
-        return Object.assign(acc, {
-          [monthIso]: {
-            ...month,
-            [iso]: modifiers,
-          },
-        });
+        acc[monthIso] = {
+          ...month,
+          [iso]: modifiers,
+        };
+
+        return acc;
       }, updatedDaysAfterAddition);
     } else {
       const monthIso = toISOMonthString(day);
@@ -550,12 +551,9 @@ export default class DayPickerSingleDateController extends React.PureComponent {
 
       const modifiers = new Set(month[iso]);
       modifiers.add(modifier);
-      updatedDaysAfterAddition = {
-        ...updatedDaysAfterAddition,
-        [monthIso]: {
-          ...month,
-          [iso]: modifiers,
-        },
+      updatedDaysAfterAddition[monthIso] = {
+        ...month,
+        [iso]: modifiers,
       };
     }
 
@@ -590,12 +588,11 @@ export default class DayPickerSingleDateController extends React.PureComponent {
         const month = updatedDays[monthIso] || visibleDays[monthIso];
         const modifiers = new Set(month[iso]);
         modifiers.delete(modifier);
-        return Object.assign(acc, {
-          [monthIso]: {
-            ...month,
-            [iso]: modifiers,
-          },
-        });
+        acc[monthIso] = {
+          ...month,
+          [iso]: modifiers,
+        };
+        return acc;
       }, updatedDaysAfterDeletion);
     } else {
       const monthIso = toISOMonthString(day);
@@ -603,12 +600,9 @@ export default class DayPickerSingleDateController extends React.PureComponent {
 
       const modifiers = new Set(month[iso]);
       modifiers.delete(modifier);
-      updatedDaysAfterDeletion = {
-        ...updatedDaysAfterDeletion,
-        [monthIso]: {
-          ...month,
-          [iso]: modifiers,
-        },
+      updatedDaysAfterDeletion[monthIso] = {
+        ...month,
+        [iso]: modifiers,
       };
     }