While profiling DayRangePickerController, I noticed that we spend a
significant amount of time generating modifiers whenever anything is
changed. A good amount of this cost comes from these reducers that
unnecessarily create new objects and spread into them on every
iteration. We can make this faster by mutating the accumulator while
inside the reducer.

In my profiling, this reduces the speed of deleteModifier from ~28ms to
~20ms, and I think it will reduce the memory usage, which should further
help reduce garbage collection times.

---

Mutate updatedDaysAfterDeletion in deleteModifier/addModifier

We create this object in this function and then return it at the end, so
it should be safe to just mutate it while we are in this function
instead of rebuilding it immediately again.

This reduces the amount of time spent in deleteModifier from ~20ms to
~16ms.

---

Avoid creating new Sets when deleting/adding modifiers

I noticed that we always create a new set and then delete from it even
if the modifier is not in the set, or create a new set and then add it
even it the modifier is already in the set. This optimization reduces
the cost of DayPickerRangeController#componentWillReceiveProps from
~16ms to ~12ms.

Before:
![image](https://user-images.githubusercontent.com/195534/58668351-75febe80-82ed-11e9-9f0d-091f25eae995.png)

After:
![image](https://user-images.githubusercontent.com/195534/58668379-8adb5200-82ed-11e9-810d-29f82d16fe14.png)