@lencioni 2019-05-30
Alright friends, tests are passing and I think this is ready for review!