@lencioni 2019-05-31
Yeah, it seems that there's a fair amount of duplicated code here, so abstracting something might make sense.